--- src/actions/FilterActions.js
@@ -1,8 +1,0 @@
-import { CHANGE_PET_FILTER } from './types';
-
-const changePetFilterParams = params => ({
-  type: CHANGE_PET_FILTER,
-  filter: params,
-});
-
-export default changePetFilterParams;
