+++ src/reducers/PetReducer.js
@@ -1,5 +1,5 @@
 import {
-  GET_PETS_LIST, PREVIOUS_PETS_LIST, NEXT_PETS_LIST, GET_PET_INFO,
+  GET_PETS_LIST, PREVIOUS_PETS_LIST, NEXT_PETS_LIST, GET_PET_INFO, CHANGE_PET_FILTER,
 } from '../actions/types';
 
 const initialState = {};
@@ -12,6 +12,10 @@ const petReducer = (state = initialState, action) => {
     case GET_PET_INFO:
       return {
         ...state,
+        ...action.payload,
+      };
+    case CHANGE_PET_FILTER:
+      return {
         ...action.payload,
       };
     default:
